patching/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws-inspector
boto3
